updated script for Tiii massage clinic evaluation
This is an updated mysql script that is running phpMyAdmin online database hosting. Please try to connect it with your web application, at the part of header, there is a connection string that must be used in your connection of your web program if you need to connect this online database sql233342, thanks.